[#10] refactor: CommentListReq, Res의 패키지 재수정
