--- src/main/java/com/b3/ddarelro/domain/comment/dto/CommentListReq.java
@@ -1,5 +1,0 @@
-package com.b3.ddarelro.domain.comment.dto;
-
-public record CommentListReq(Long cardId) {
-
-}
