+++ src/main/java/com/b3/ddarelro/domain/comment/dto/request/CommentListReq.java
@@ -1,0 +1,5 @@
+package com.b3.ddarelro.domain.comment.dto.request;
+
+public record CommentListReq(Long cardId) {
+
+}
