+++ src/main/java/com/b3/ddarelro/domain/comment/dto/response/CommentListRes.java
@@ -1,4 +1,4 @@
-package com.b3.ddarelro.domain.comment.dto;
+package com.b3.ddarelro.domain.comment.dto.response;
 
 import lombok.Builder;
